+++ api_collector/route/route_collector.py
@@ -1,5 +1,5 @@
-from api_data.air_tickets.air_tickets_api import AirTicketsApi
-from api_data.hotels.hotel_api import HotelApi
+from api_collector.air_tickets.air_tickets_api import AirTicketsApi
+from api_collector.hotels.hotel_api import HotelApi
 
 
 class RouteCollector:
